+++ ui/voice/src/interfaces.ts
@@ -1,30 +1,5 @@
 import { Api as CgApi } from 'chessground/api';
 import * as cg from 'chessground/types';
-import { Prop } from 'common';
-
-export type MoveHandler = (fen: Fen, api: CgApi, yourMove?: boolean) => void;
-
-export interface InputOpts {
-  input?: HTMLInputElement;
-  ctrl: VoiceMove;
-}
-
-export interface SubmitOpts {
-  isTrusted: boolean;
-  force?: boolean;
-  yourMove?: boolean;
-}
-
-export type Submit = (v: string, submitOpts: SubmitOpts) => void;
-
-export interface VoiceMove {
-  available(): [string, string][];
-  arrogance(conf?: number): number;
-  arrowColors(enabled?: boolean): boolean;
-  update(step: { fen: string }, yourMove?: boolean): void;
-  modalOpen: Prop<boolean>;
-  root: RootCtrl;
-}
 
 export interface CrazyPocket {
   [role: string]: number;
